+++ changelog.md
@@ -1,6 +1,7 @@
 # Changelog
 
-- * 63b24b5 - 2018-08-22: Pre-Release v9.3.58.dev1 
+- * 6f070ab - 2018-08-23: Switch TorrentLeech provider to login via cookies due to reCAPTCHA 
+- * 2f40c76 - 2018-08-22: Pre-Release v9.3.58.dev1 
 - * 9010755 - 2018-08-22: Fixed issue #256 - unable to search BTN 
 - * 5f0e3bf - 2018-08-20: Fixed issue #253 - Provider Option Missing Cookie Field 
 - * e5b5706 - 2018-08-19: Release v9.3.57